ipp-proxy: Add url-mapper
url-mapper is a "URL rewriting filter" for use with
`url_rewrite_program` in `squid.conf`. It can support printers that are
picky about the URL path sent in an IPP request (mine is not), or it
could be extended to look up credentials (authentication tokens) in a
database or similar.